Channel PlugValueWidgets : Avoid unnecessary updates
We were using ImagePlugs as our auxiliary plugs as it was a convenient way of depending on both the `viewNames` and `channelNames` plugs. But this meant an update was triggered every time the `ImagePlug["channelData"]` plug was dirtied too, which was basically continuous when viewing an interactive render. This resulted in the channel selector in the ImageView flickering constantly, and the dropdown menu sometimes being empty because an update was pending. This is solved by declaring our dependencies more accurately, returning just the `viewNames` and `channelNames` plugs.

I've also deliberately removed the attempt to handle the lack of an ImagePlug in `_auxiliaryPlugs()` for RGBChannelsPlugValueWidget and ChannelPlugValueWidget, as there was no handling for the resulting empty lists in the subsequent `_valuesForUpdate()` methods. This matches the situation before we introduced asynchronous updates - we have always required the ImagePlug to exist and thrown if it doesn't.

Also moved error indicator code above the early out in `RGBAChannelsPlugValueWidget._updateFromValues()`, otherwise we'd never show the error state